backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.9
argon2-cffi==23.1.0
python-jose==3.3.0
cryptography==43.0.1
python-multipart==0.0.9
redis==5.0.8
httpx==0.27.2
